Just needs comments
Finished the background, gave the nav bar a more fancy look, gave the heading 1 elements a text shadow, then gave the portrait some filters and rotated it, also put progressive enhancements, then put a border on the footer. Went through and fixed things that made the page not look right.